Fix for compatibility
