+++ wvlet-log/js/src/main/scala/wvlet/log/LogUtil.scala
@@ -1,0 +1,17 @@
+package wvlet.log
+
+/**
+  *
+  */
+object LogUtil {
+  def getSuccinctLoggerName(cl: Class[_]): String = {
+    val name =cl.getName
+    if (name.endsWith("$")) {
+      // Remove trailing $ of Scala Object name
+      name.substring(0, name.length - 1)
+    }
+    else {
+      name
+    }
+  }
+}
