+++ wvlet-log/jvm/src/test/scala/wvlet/log/JVMSpec.scala
